#122: support for memory output formatter and SIMULATION output types added, changed rc convergence gap property from float to double, minor refactoring

+++ src/main/java/org/goplanit/output/property/RouteChoiceConvergenceGapOutputProperty.java
@@ -24,7 +24,7 @@ public final class RouteChoiceConvergenceGapOutputProperty extends OutputPropert
 
   @Override
   public DataType getDataType() {
-    return DataType.FLOAT;
+    return DataType.DOUBLE;
   }
 
   @Override
